transformed 'DEVELOPER_MODE' into 'environment', moved error-handling init into apps, re-worked the boot-up process to be more flexible and require less global variables between scripts
It felt extremely wrong, when I started to document the old master.php and had
to write 'the interface to this class consists of the following global
variables: $loader should contain ...'. As the core isn't really an app itself,
I think it's okay to not have a boot.php here.

As per Zozi's wish, we'll keep the SLY_* constants for now. But hopefully they
will be gone sometime.

--HG--
branch : 0.8
rename : loader.php => autoload.php