use a different approach to selecting the max revision when finding articles and categories
The original approach was buggy because the GROUP BY clause did not select the
first row for each article, but the last (by spec, as there is no defined order
for the grouping to work). The new query is stable in that regard and avoids an
explicit derived table.
The additional indexes just help common queries (since we look through the table
twice per query, we should make sure it goes reasonably fast).